[Subtask]: Modify the optimizer to support obtaining tasks from each AMS node for processing. #3928
[Subtask]: Modify the optimizer to support obtaining tasks from each AMS node for processing. #3928

[Subtask]: Optimize the logic for retrieving the AMS list from ZooKeeper in master-slave mode. #3928

This addresses the conflict issues with the latest main branch and introduces a new solution for storing allocation information based on a database.

This addresses the conflict issues with the latest main branch and introduces a new solution for storing allocation information based on a database.

Troubleshooting unit test failures.